+++ src/Symfony/Component/Messenger/Bridge/MongoDb/CHANGELOG.md
@@ -5,3 +5,4 @@ CHANGELOG
 ---
 
  * Introduce the MongoDB Messenger bridge
+ * Add Change Streams to wake idle workers up as soon as a message is inserted, with the `wait_time` option (0 disables it)